extract develop util functions into different file
