+++ src/pages/develop.astro
@@ -2,7 +2,7 @@
 import Layout from "../layouts/Layout.astro";
 ---
 
-<script src="/develop.js" defer></script>
+<script type="module" src="/develop.js" defer></script>
 
 <Layout title="Develop - OrnitheMC">
   <main class="flex flex-col gap-4">
